requirements.txt
segno
cairosvg
pillow
